+++ src/main/java/com/fypgrading/adminservice/service/dto/NotificationDTO.java
@@ -14,9 +14,9 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 public class NotificationDTO {
 
-    private Integer id;
+    private Long id;
 
-    private Integer teamId;
+    private Long teamId;
 
     private AssessmentEnum assessment;
